Updated h_TableMemoryManagemet.hpp added 'explicit' keyword to 'Table' constructor.
Explanation with example:
Adding the 'explicit' keyword to constructor: 'explicit Table(int num);'
it prevent the compiler from implicitly converting an 'int' to a 'Table' when a 'Table' is expected as a
parametert.
Example of this:
	 void foo(const Table& t);
foo(5); // Not allowed if Table(int) is explicit
foo(Table(5)); // Allowed
So, 'explicit' ensures that you (and others) must always construct a 'Table' object direclty, making code the code safer and more predictable.